fix db migrations for production

+++ db/migrate/20151206161455_drop_users.rb
@@ -1,5 +1,7 @@
 class DropUsers < ActiveRecord::Migration
   def change
-    drop_table(:users)
+    if table_exists?(:users)
+      drop_table(:users)
+    end
   end
 end
